+++ src/store/action.ts
@@ -3,6 +3,7 @@ import { Offers } from '../types/offer';
 import { AuthorizationStatus } from '../components/const';
 import { FullOffer } from '../types/full-offer';
 import { Reviews } from '../types/reviews';
+import { UserData } from '../types/user-data';
 
 export const changeCurrentCity = createAction<{currentCity:string}>('cities/currentCity');
 export const changeCurrentOffers = createAction('cities/currentOffers');
@@ -10,6 +11,7 @@ export const loadOffers = createAction<Offers>('cities/loadOffers');
 export const loadOffer = createAction<FullOffer>('cities/loadOffer');
 export const loadReviews = createAction<Reviews[]>('cities/loadReviews');
 export const addReview = createAction<Reviews>('cities/addReview');
+export const getUserData = createAction<UserData>('getUserData');
 export const loadNearbyOffers = createAction<Offers>('cities/loadNearbyOffers');
 export const requireAuthorization = createAction<AuthorizationStatus>('user/requireAuthorization');
 export const setError = createAction<string | null>('game/setError');
