+++ src/types/user-data.ts
@@ -1,5 +1,5 @@
 export type UserData = {
-  id: number;
   email: string;
   token: string;
+  avatarUrl: string;
 };
